move prom
