+++ beacon-chain/node/BUILD.bazel
@@ -37,6 +37,7 @@ go_library(
         "//beacon-chain/sync/initial-sync:go_default_library",
         "//cmd/beacon-chain/flags:go_default_library",
         "//monitoring/backup:go_default_library",
+        "//monitoring/prometheus:go_default_library",
         "//monitoring/tracing:go_default_library",
         "//shared:go_default_library",
         "//shared/cmd:go_default_library",
@@ -46,7 +47,6 @@ go_library(
         "//shared/gateway:go_default_library",
         "//shared/params:go_default_library",
         "//shared/prereq:go_default_library",
-        "//shared/prometheus:go_default_library",
         "//shared/sliceutil:go_default_library",
         "//shared/version:go_default_library",
         "@com_github_ethereum_go_ethereum//common:go_default_library",
