+++ validator/node/BUILD.bazel
@@ -30,6 +30,7 @@ go_library(
     deps = [
         "//cmd/validator/flags:go_default_library",
         "//monitoring/backup:go_default_library",
+        "//monitoring/prometheus:go_default_library",
         "//monitoring/tracing:go_default_library",
         "//proto/prysm/v1alpha1:go_default_library",
         "//proto/prysm/v1alpha1/validator-client:go_default_library",
@@ -42,7 +43,6 @@ go_library(
         "//shared/gateway:go_default_library",
         "//shared/params:go_default_library",
         "//shared/prereq:go_default_library",
-        "//shared/prometheus:go_default_library",
         "//shared/version:go_default_library",
         "//validator/accounts/iface:go_default_library",
         "//validator/accounts/wallet:go_default_library",
